backend/requirements.txt
fastapi
uvicorn
scikit-learn
vaderSentiment
appdirs==1.4.4
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cssselect==1.3.0
curl_cffi==0.11.3
cycler==0.12.1
fake-useragent==2.2.0
feedparser==6.0.11
fonttools==4.58.4
frozendict==2.4.6
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
multitasking==0.0.11
numpy==1.24.4
packaging==25.0
pandas==2.3.0
parse==1.20.2
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
protobuf==3.20.3
pycparser==2.22
pyee==11.1.1
pyparsing==3.2.3
pyppeteer==0.0.25
pyquery==2.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-html==0.10.0
seaborn==0.13.2
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
w3lib==2.3.1
websockets==15.0.1
yahoo-fin==0.8.9.1
yfinance==0.2.63
zipp==3.23.0
passlib[bcrypt]
python-multipart
textblob
tensorflow-cpu
pandas-ta
